adjust database to migrate

+++ api/models/synth.py
@@ -10,7 +10,7 @@ class Synth(db.Model):
     model = db.Column(db.String(100))
     finish = db.Column(db.String(100))
     year = db.Column(db.String(4))
-    number_of_keys = db.Column(db.Integer(5))
+    number_of_keys = db.Column(db.Integer)
     description = db.Column(db.String(250))
     condition = db.Column(db.String(100))
     is_working = db.Column(db.Boolean, default=True, nullable=False)
